pyproject.toml
[project]
name = "agente-pd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.12",
    "langchain-anthropic>=1.3.5",
    "langgraph>=1.1.2",
    "pypdf>=6.8.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "playwright>=1.49.0",
    "llama-index>=0.14.16",
    "llama-index-embeddings-huggingface>=0.7.0",
    "llama-index-llms-anthropic>=0.11.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.20",
    "langchain-google-genai>=4.2.1",
]
